requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
builtwith==1.3.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
future==0.18.2
idna==2.10
instaloader==4.4.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pefile==2019.4.18
pipenv==2020.6.2
psutil==5.7.0
pycparser==2.20
PyInstaller==3.6
pylint==2.5.3
pynput==1.6.8
pyTelegramBotAPI==3.7.1
python-decouple==3.3
python-telegram-bot==12.8
pywin32-ctypes==0.2.0
requests==2.24.0
scapy==2.4.3
six==1.15.0
telegram-send==0.25
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
virtualenv==20.0.26
virtualenv-clone==0.5.4
wrapt==1.12.1
